--- examples/acme_examples/.gitignore
@@ -1,1 +1,0 @@
-test*
